Split Http.scala file into multiple files
Finally RequestHeader and Request are in their own files! No code was
changed, except I removed GlobalStateHttpConfiguration after inlining
its httpConfiguration method.
